+++ H6Game/H6Game.Test/TestDistributed/TestHandler.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace TestDistributed
 {
-    [MessageCMD((int)MessageCMD.TestCMD1)]
+    [MessageCMD(MessageCMD.TestCMD1)]
     public class TestHandler : AHandler<string>
     {
         protected override void Dispatcher(string response, int messageId)
